requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
anndata=0.10.1=pypi_0
array-api-compat=1.4=pypi_0
blas=1.0=mkl
boost=1.74.0=py39h5472131_5
boost-cpp=1.74.0=h75c5d50_8
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.3.11=h06a4308_0
cairo=1.16.0=hb05425b_5
captum=0.7.0=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=3.0.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=42.0.5=py39hdda0065_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.2.140=0
cuda-runtime=12.1.0=0
cudatoolkit=11.8.0=h4ba93d1_13
cycler=0.11.0=pyhd3eb1b0_0
exceptiongroup=1.1.3=pypi_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.9.0=py39h06a4308_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
greenlet=2.0.1=py39h6a678d5_0
h5py=3.9.0=pypi_0
icu=70.1=h27087fc_0
idna=3.4=py39h06a4308_0
igraph=0.10.8=pypi_0
imbalanced-learn=0.11.0=pypi_0
imblearn=0.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=py39h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py39h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_2
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hcd5def8_4
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=13.2.0=h7e041cc_2
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h22db469_4
libzlib=1.2.13=hd590300_5
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.7.2=py39h1128e8f_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
munkres=1.1.4=py_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numba=0.58.0=pypi_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.24.3=py39h14f4228_0
numpy-base=1.24.3=py39h31eccc5_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.0=h4ab18f5_3
opentsne=1.0.1=py39h85c637f_2
packaging=23.1=py39h06a4308_0
pandas=2.0.3=py39h1128e8f_0
patsy=0.5.3=pypi_0
pcre=8.45=h295c915_0
pcre2=10.42=hebb0a14_0
pillow=10.0.1=py39ha6cbd5a_0
pip=23.2.1=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=3.10.0=py39h06a4308_0
pooch=1.7.0=py39h06a4308_0
psutil=5.9.5=pypi_0
pycairo=1.23.0=py39hd1222b9_0
pycparser=2.21=pyhd3eb1b0_0
pynndescent=0.5.10=pypi_0
pyopenssl=24.0.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=2.1.0=py3.9_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
rdkit=2022.03.5=py39h89e00b9_0
readline=8.2=h5eee18b_0
reportlab=3.5.67=py39hfdd840d_1
requests=2.31.0=py39h06a4308_0
scanpy=1.9.5=pypi_0
scikit-learn=1.3.1=pypi_0
scipy=1.10.1=py39h14f4228_0
seaborn=0.13.0=pypi_0
session-info=1.0.0=pypi_0
setuptools=68.0.0=py39h06a4308_0
shap=0.45.1=cuda118py39h666a580_0
six=1.16.0=pyhd3eb1b0_1
slicer=0.0.8=pyhd8ed1ab_0
sqlalchemy=2.0.21=py39h5eee18b_0
sqlite=3.41.2=h5eee18b_0
statsmodels=0.14.0=pypi_0
stdlib-list=0.9.0=pypi_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.10.0=pypi_0
texttable=1.7.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch-cluster=1.6.2+pt21cu121=pypi_0
torch-geometric=2.3.1=pypi_0
torch-scatter=2.1.2+pt21cu121=pypi_0
torch-sparse=0.6.18+pt21cu121=pypi_0
torch-spline-conv=1.2.2+pt21cu121=pypi_0
torchaudio=2.1.0=py39_cu121
torchtriton=2.1.0=py39
torchvision=0.16.0=py39_cu121
tqdm=4.66.1=pypi_0
typing-extensions=4.7.1=py39h06a4308_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
umap-learn=0.5.4=pypi_0
urllib3=1.26.16=py39h06a4308_0
wheel=0.41.2=py39h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
